requirement.txt
requests
python-dotenv
langchain-community
langchain-openai
langchain
langgraph
openpyxl 
pandas
numpy